Pre-generate access tables for zoned random distribution
If we have a lot of zones, then the looping to find and sum our
correct one eats some CPU cycles. On a null workload with 10 zones,
it was about a 20% reduction in performance.

So cache the state information, by pre-generating the tables. This
makes it so that there's no difference in performance between running
without zones, or with tens of zones.

Signed-off-by: Jens Axboe <axboe@fb.com>